requirements_deploy.txt
# Core Telegram Bot Framework
aiogram==3.22.0
aiohttp==3.11.11
aiodns==3.2.0
aiosignal==1.3.2
frozenlist==1.5.0
multidict==6.1.0
yarl==1.18.3

# Database (SQLite)
# sqlite3 - встроен в Python, установка не требуется

# Environment and Utils
python-dotenv==1.2.1
requests==2.32.5
pytz==2025.2
certifi==2024.12.14
charset-normalizer==3.4.1
idna==3.12
urllib3==2.3.0

# Excel Reports and Charts
openpyxl==3.1.5
et-xmlfile==2.0.0
matplotlib==3.10.7
pillow==12.0.0
numpy==2.3.5
contourpy==1.3.3
cycler==0.12.1
fonttools==4.60.1
kiwisolver==1.4.9
packaging==25.0
pyparsing==3.2.5
python-dateutil==2.9.0.post0
six==1.17.0

# Optional (Flask not used by bot but installed)
flask==3.1.0
Werkzeug==3.1.3
Jinja2==3.1.5
MarkupSafe==3.0.2
itsdangerous==2.2.0
click==8.1.8
blinker==1.9.0
